+++ packages/extension-invisible-characters/src/plugin/reducers.ts
@@ -1,0 +1,14 @@
+import { PluginState } from '../types.js'
+
+export const stateReducer = (
+  state: PluginState,
+  visible: boolean,
+): PluginState => {
+  if (visible === undefined) {
+    return state
+  }
+
+  return { ...state, visible }
+}
+
+export default stateReducer