+++ packages/extension-youtube/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 2.22.3
+
+## 2.22.2
+
+## 2.22.1
+
+### Patch Changes
+
+- e465825: Update YouTube URL regex pattern
+
+## 2.22.0
+
+## 2.14.1
+
 ## 2.14.0
 
 ## 2.13.0