+++ packages/vue-3/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Change Log
 
+## 2.22.3
+
+### Patch Changes
+
+- @tiptap/extension-bubble-menu@2.22.3
+- @tiptap/extension-floating-menu@2.22.3
+
+## 2.22.2
+
+### Patch Changes
+
+- @tiptap/extension-bubble-menu@2.22.2
+- @tiptap/extension-floating-menu@2.22.2
+
+## 2.22.1
+
+### Patch Changes
+
+- @tiptap/extension-bubble-menu@2.22.1
+- @tiptap/extension-floating-menu@2.22.1
+
+## 2.22.0
+
+### Patch Changes
+
+- @tiptap/extension-bubble-menu@2.22.0
+- @tiptap/extension-floating-menu@2.22.0
+
+## 2.14.1
+
+### Patch Changes
+
+- @tiptap/extension-bubble-menu@2.14.1
+- @tiptap/extension-floating-menu@2.14.1
+
 ## 2.14.0
 
 ### Patch Changes